API

Modules

Types and constants

UpROOT.OpaqueObject

UpROOT.OpaqueObjectArray

UpROOT.TBranch

UpROOT.TDirectory

UpROOT.TFile

UpROOT.TTree

Functions and macros

UpROOT.pyobj

UpROOT.testdatadir

Documentation

UpROOT.OpaqueObject — Type

UpROOT.OpaqueObject{tp,B<:AbstractVector{UInt8}}

An opaque object of symbolic type tp. tp must be a Symbol. Use the data field to access the byte representation.

UpROOT.OpaqueObjectArray — Type

OpaqueObjectArray{tp,N,BA<:AbstractArray{<:AbstractBlob,N}}

An array of [UpROOT.OpaqueObject]s.

UpROOT.TBranch — Type

TBranch <: AbstractVector{Any}

UpROOT.TBranch is a wrapper around Python objects with mix-in uproot.tree.TBranchMethods. It behaves like a Julia AbstractVector. Ff the branch has children, it also behaves like a Tables.Table (with column access).

TBranch is an on-disk data structure, iteration over single elements is very inefficient.

Limitations: Write access is not implemented yet.

UpROOT.TDirectory — Type

TDirectory

UpROOT.TDirectory is a wrapper around Python objects of type uproot.rootio.ROOTDirectory.

TDirectory behaves similar to a Julia Dict, it supports the functions keys, length and getindex.

Limitations: Write access is not implemented yet.

TFile is defined as an alias for TDirectory in UpROOT.jl.

UpROOT.TFile — Type

TFile = TDirectory

The uproot Python package doesn't use separate types for the ROOT classes TFile and TDirectoy - so in UpROOT.jl, so in UpROOT.jl, TFile is just defined as an alias for TDirectory.

Constructors:

* `TFile(filename::AbstractString)::TFile`

Use TFile(filename) to open files instead of TDirectory(filename) for increased clarity in your code.

UpROOT.TTree — Type

TTree <: AbstractVector{Any}

UpROOT.TTree is a wrapper around Python objects with mix-in uproot.tree.TTreeMethods. It behaves like a Julia AbstractVector and Tables.Table (with column access).

TTree is an on-disk data structure, iteration over single elements is very inefficient.

Even though TTree is semantically a vector over the tree entries, a two dimensional getindex(tree::TTree, entries, branches) is defined for convenience. So you may use, e.g.:

tree[1:5, (:branch1, :branch2)]
tree[:, [:branch1, :branch2]]
tree[[1,4,7,10], ["branch1", "branch2"]]

Limitations: Write access is not implemented yet.

UpROOT.pyobj — Method

UpROOT.pyobj(x)::PyObject

Get the python object wrapped by Julia object x. x may be a TFile/TDirectory, TDirectory or TBranch.

UpROOT.testdatadir — Method

UpROOT.testdatadir()::String

Path to some test data installed by UpROOT.jl for testing and experimentation purposes.

Use UpROOT.testfiles to get Dict of all test files.