API

API

Types

UpROOT.TBranch

UpROOT.TDirectory

UpROOT.TFile

UpROOT.TTree

Functions

UpROOT.pyobj

UpROOT.testdatadir

Documentation

UpROOT.TBranch — Type.

TBranch <: AbstractVector{Any}

UpROOT.TBranch is a wrapper around Python objects with mix-in uproot.tree.TBranchMethods. It behaves like a Julia AbstractVector. Ff the branch has children, it also behaves like a Tables.Table (with column access).

Limitations: Write access is not implemented yet.

source

UpROOT.TDirectory — Type.

TDirectory

UpROOT.TDirectory is a wrapper around Python objects of type uproot.rootio.ROOTDirectory.

TDirectory behaves similar to a Julia Dict, it supports the functions keys, length and getindex.

Limitations: Write access is not implemented yet.

TFile is defined as an alias for TDirectory in UpROOT.jl.

source

UpROOT.TFile — Type.

TFile = TDirectory

The uproot Python package doesn't use separate types for the ROOT classes TFile and TDirectoy - so in UpROOT.jl, so in UpROOT.jl, TFile is just defined as an alias for TDirectory.

Constructors:

* `TFile(filename::AbstractString)::TFile`

Use TFile(filename) to open files instead of TDirectory(filename) for increased clarity in your code.

source

UpROOT.TTree — Type.

TTree <: AbstractVector{Any}

UpROOT.TTree is a wrapper around Python objects with mix-in uproot.tree.TTreeMethods. It behaves like a Julia AbstractVector and Tables.Table (with column access).

Limitations: Write access is not implemented yet.

source

UpROOT.pyobj — Method.

UpROOT.pyobj(x)::PyObject

Get the python object wrapped by Julia object x. x may be a TFile/TDirectory, TDirectory or TBranch.

source

UpROOT.testdatadir — Method.

UpROOT.testdatadir()::String

Path to some test data installed by UpROOT.jl for testing and experimentation purposes.

Use UpROOT.testfiles to get Dict of all test files.

source